feat: Take page temporary down (show back soon)
Because Swisscom shut down their service as problems arose.
